+++ projekt/update_table.sql
@@ -1,0 +1,1 @@
+UPDATE CUSTOMERS_400049 SET SecretCode= substr(md5(random()::text), 0,10);